dot.login:
Finally transform the "Don't login as root..." message to make it
clear that `su' is meant to be a command.  Will save us a lot of
questions about the user named `su'.  Make the message magically
disappear if the user did an ``su - root'', since it might be a bit
silly to ask him to perform an `su'...

dot.cshrc:

Remove the no-op `-g' options from the ls aliases, and replace them by
`-o'.  This way, if root does an `ls -l', he will see the immutable
flag and (hopefully) not be too surprised about the "Permission
denied".