Please all welcome the long-awaited upgrade from our ancient xntpd 3.4f to a brand new and shiny ntpd 4.0.98f.
I got tired of waiting for 4.1.0 and there is the feature freeze deadline
so here it is. This is the contrib/ part of the upgrade. The Makefile glue
will be added very soon in usr.sbin.

It builds and runs on both i386 and alpha (Thanks Peter!).

The bad news is that manpages no longer exist, everything is in HTML. I'll
commit the text version of each HTML file in /usr/share/doc/ntp soon to have
at least the help files w/o needing to get the entire contrib/ntp tree.

I'll commit FREEBSD-Xlist as soon as I can skip over $FreeBSD$ checks...

Reviewed by:	peter, obrien
Pushed by:	phk